fix: context file and alert message
