+++ thyrel-web/src/utils/utils.ts
@@ -1,4 +1,2 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const callAll = (...fns: (any | undefined)[]) => (...args: any) =>
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   fns.forEach((fn: any) => fn && fn(...args));
